requirements.txt
pyyaml>=3.12
pydruid>=0.3.1
pandas>=0.22.0
matplotlib[plots]>=2.1.1
numpy[plots]>=1.13.3